Extract random number function and content for home page to seperate files
