+++ src/App.tsx
@@ -2,10 +2,9 @@ import Home from "./pages/Home";
 
 function App() {
   return (
-    <>
-      <h1 className="text-3xl font-bold underline text-red-300">Main Page</h1>
+    <main>
       <Home />
-    </>
+    </main>
   );
 }
