+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export const getRandomNumber = (arr: string[]) => {
+  return Math.floor(Math.random() * arr.length);
+};
